requirements.txt
torch>=2.2.0
tiktoken>=0.5.0
datasets>=2.14.0
pyyaml>=6.0
wandb>=0.16.0
